requirements.txt
icrawler>=0.6.6
requests>=2.31.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66.0
